requirements.txt
# ニュース要約掲示板アプリ - 必要なライブラリ
# News Summary Display App - Required Libraries

# OpenAI API クライアント
openai>=1.0.0

# ニュース API 取得用
requests>=2.28.0

# LCD 制御用ライブラリ
RPLCD>=1.3.0

# I2C 通信用
smbus2>=0.4.1

# 環境変数管理
python-dotenv>=1.0.0